+++ src/vs/workbench/contrib/terminal/browser/terminalVoiceActions.ts
@@ -1,0 +1,36 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { localize2 } from '../../../../nls.js';
+import { ContextKeyExpr } from '../../../../platform/contextkey/common/contextkey.js';
+import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
+import { HasSpeechProvider } from '../../speech/common/speechService.js';
+import { TerminalCommandId } from '../common/terminal.js';
+import { registerActiveInstanceAction, sharedWhenClause } from './terminalActions.js';
+import { TerminalVoiceSession } from './terminalVoice.js';
+
+export function registerTerminalVoiceActions() {
+	registerActiveInstanceAction({
+		id: TerminalCommandId.StartVoice,
+		title: localize2('workbench.action.terminal.startDictation', "Start Dictation in Terminal"),
+		precondition: ContextKeyExpr.and(HasSpeechProvider, sharedWhenClause.terminalAvailable),
+		f1: true,
+		run: (activeInstance, c, accessor) => {
+			const instantiationService = accessor.get(IInstantiationService);
+			TerminalVoiceSession.getInstance(instantiationService).start();
+		}
+	});
+
+	registerActiveInstanceAction({
+		id: TerminalCommandId.StopVoice,
+		title: localize2('workbench.action.terminal.stopDictation', "Stop Dictation in Terminal"),
+		precondition: ContextKeyExpr.and(HasSpeechProvider, sharedWhenClause.terminalAvailable),
+		f1: true,
+		run: (activeInstance, c, accessor) => {
+			const instantiationService = accessor.get(IInstantiationService);
+			TerminalVoiceSession.getInstance(instantiationService).stop(true);
+		}
+	});
+}
